Changes for booting.
OK, at this point,
- kernel can not start graphics without this code
- this code can not start graphics without kernel

What a mess.

We know that some of the programming in here is wrong for 32 bits direct
kernel -- the VBIOS we got this code from is just doing the wrong stuff.

Until wednesday, when we hope to have a better vbios to test, we're going to make a few changes.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>